refactor(tests): remove XCTest dependency from BaseChatTestSupport
Moves `StopGenerationContractTests` helper out of `BaseChatTestSupport`
(its only XCTest consumer) into `Tests/BaseChatInferenceTests/` where its
sole caller lives. Deletes the inlined `Sources/fuzz-chat/OllamaDiscovery.swift`
and rewires `fuzz-chat` to reuse `HardwareRequirements.findOllamaModel(...)`
from `BaseChatTestSupport`. Adds `listOllamaModels()` and
`findOllamaModel(nameContains:)` helpers to preserve the CLI's `--model`
substring-hint behaviour.

`BaseChatTestSupport` is now XCTest-free, so non-test consumers
(executables like `fuzz-chat`) can depend on it without dragging
`libXCTestSwiftSupport.dylib` into their link graph.

Closes #495
